move fakeLMA to io
